requirements.txt
transformers==4.6.0
PyYAML==5.3.*
pandas
scikit-learn==0.24
torch<=1.6
numpy<=1.20.0
scipy==1.5.
tensorboard==2.2.0
pytorch-lightning==1.2.3
pytorch-nlp==0.5.0
optuna
plotly
sentencepiece
